FEAT: start swipe position determines if answer was correct
user can swipe from the upper half of the flashcard if he didn't answer correctly, changing the border of the card to red and adding the card to the end of the array.

if the user swipes from the lower half of the flashcard the border will be set green and the card won't be added to the end.